+++ Demo/SharedScript.ps1
@@ -6,11 +6,11 @@
 
 <#
 .Synopsis
-	Formats error information for testing.
+	Formats error information.
 
 .Description
-	Code "$$ | Out-String" does not work in some hosts with strict mode. It
-	works but adds errors like "Width is not found" and gets empty text.
+	It replaces formatting by Out-String which does not work in strict mode
+	with some hosts due to internal errors like "property not found".
 #>
 function Format-Error($$) {
 	@"
@@ -18,4 +18,25 @@ $$
 $($$.InvocationInfo.PositionMessage.Trim().Replace("`n", "`r`n"))
 + $($$.CategoryInfo)
 "@
+}
+
+<#
+.Synopsis
+	Invokes a build with an issue and comparer the error with a pattern.
+#>
+function Test-Issue([Parameter()]$Task, $File, $ExpectedPattern) {
+	$message = ''
+	try { Invoke-Build $Task $File }
+	catch { $message = Format-Error $_ }
+	Write-BuildText Magenta $message
+	if ($message -notlike $ExpectedPattern) {
+		Write-Error -ErrorAction Stop @"
+Expected pattern [
+$ExpectedPattern
+]
+Actual error [
+$message
+]
+"@
+	}
 }